pyproject.toml
[tool.poetry]
name = "drf"
version = "0.1.0"
description = ""
authors = ["lilimik <kazakovlim@inbox.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.1"
djangorestframework = "^3.12.4"
gunicorn = "^20.1.0"
djangorestframework-simplejwt = "^4.7.2"
psycopg2 = "^2.9.1"
python-dotenv = "^0.18.0"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
